+++ src/transaction/mod.rs
@@ -12,7 +12,7 @@ use crate::{
     },
     utils::Address,
 };
-pub use declare::DeclareDeprecated;
+pub use declare_deprecated::DeclareDeprecated;
 pub use declare_v2::DeclareV2;
 pub use deploy::Deploy;
 pub use deploy_account::DeployAccount;
@@ -20,7 +20,7 @@ use error::TransactionError;
 pub use invoke_function::InvokeFunction;
 pub use l1_handler::L1Handler;
 
-pub mod declare;
+pub mod declare_deprecated;
 pub mod declare_v2;
 pub mod deploy;
 pub mod deploy_account;
